separate job template and projects
Now the job templates are stored in templates/job_templates.yaml.
Technically they migh be splitted in several files withing templates folder.
It doesn't matter.

The projects are stored in сoppresponding folders product-9.0, product-9.1 and
so on. The aim is to use same templates across all projects for all milestones.

For 9.1 and 10.0 the tempest jobs are prepared.i  The time triggers for them
are empty, so the jobs will be disabled so far.

Change-Id: Ib8bb9a89983baffb48fe575a8adc0115bc736ce8