Fix snapshots updating and fuel-qa issue
After merging of cd6f3d1f0736fb20e1f5935ad557034552d94da4 commit to
fuel-qa we have a problem with deployment of environment. This patch
sets some variables in order to cancel the commit.
Also we had a problem that we tried to export variables without
ENV_INJECT jenkins plugin. This issue is also fixed.

Change-Id: I92bda2f6c8f7aa1885e81d4723087ffb4e2e2555