Take actual scheduler filters from nova.conf
By default, all scheduler filters are enabled in Tempest config.
Because of it, some Tempest tests which are expecting for specific
filters provided by Nova will fail (namely, the
api.compute.admin.test_servers_on_multinodes.ServersOnMultiNodesTest
test group needs SameHostFilter and DifferentHostFilter to be enabled),
if Nova doesn't have it.
The patch proposes to take actual list of enabled scheduler filters
from nova.conf and put it to tempest.conf instead of default 'all'
value. So, if required filter is missing on env, the appropriate test
will be skipped instead of failed.

Change-Id: I8f15963d783e3bed4c540f3062a73b528438e79b