add os faults config and update revert last script
Updated revert_last_snapshot to ignore variable SNAPSHOT_NAME stated previously
to have possibility to revert the very last snapshot even in case if deploy
failed to finish for any reasons. Mainly this is to help in case of failed OSTF
during deploy.

And added OS_FAULTS_CONIFG variable and file generation in the stepler jobs.

Change-Id: Ifd39288d97a7476c6529fb388ab9b612a9e8050e